requirements.txt
fastapi
uvicorn
python-multipart
pymongo
qdrant-client
openai
python-docx
PyPDF2
beautifulsoup4
pydantic
python-dotenv
tiktoken
